pyproject.toml
[project]
name = "video-agent-mcp"
version = "1.0.0"
description = "A comprehensive Model Context Protocol (MCP) server for AI-powered video creation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Video Agent Team"},
]
keywords = ["mcp", "ai", "video", "generation", "fal", "youtube"]

dependencies = [
    # Core MCP framework - FastMCP 2.0
    "fastmcp>=2.0.0",
    
    # FAL AI client
    "fal-client>=0.4.0",
    
    # HTTP client
    "httpx>=0.24.0",
    
    # Data validation
    "pydantic>=2.0.0",
    
    # Async file operations
    "aiofiles>=23.0.0",
    
    # Environment variables
    "python-dotenv>=1.0.0",
    
    # Type hints support
    "typing-extensions>=4.5.0",
    
    # YouTube API integration
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.0",
    "httplib2>=0.20.0",
    
    # Gemini API for video understanding
    "google-genai>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
video-agent-mcp = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"tests/*" = ["F401", "F811"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true